[backend] fix(payload): improve code
