+++ openaev-api/src/main/java/io/openaev/rest/inject/service/InjectService.java
@@ -53,8 +53,8 @@ import io.openaev.rest.injector_contract.output.InjectorContractFullOutput;
 import io.openaev.rest.security.SecurityExpression;
 import io.openaev.rest.security.SecurityExpressionHandler;
 import io.openaev.rest.tag.TagService;
-import io.openaev.rest.threat_arsenal.ThreatArsenalService;
 import io.openaev.service.*;
+import io.openaev.service.threat_arsenal.ThreatArsenalService;
 import io.openaev.utils.FilterUtilsJpa;
 import io.openaev.utils.InjectUtils;
 import io.openaev.utils.JpaUtils;
