+++ openaev-api/src/test/java/io/openaev/rest/inject/service/InjectServiceTest.java
@@ -28,13 +28,13 @@ import io.openaev.rest.inject.form.*;
 import io.openaev.rest.injector_contract.InjectorContractContentUtils;
 import io.openaev.rest.injector_contract.InjectorContractService;
 import io.openaev.rest.tag.TagService;
-import io.openaev.rest.threat_arsenal.ThreatArsenalService;
 import io.openaev.service.AssetGroupService;
 import io.openaev.service.AssetService;
 import io.openaev.service.EndpointService;
 import io.openaev.service.InjectorService;
 import io.openaev.service.TagRuleService;
 import io.openaev.service.UserService;
+import io.openaev.service.threat_arsenal.ThreatArsenalService;
 import io.openaev.utils.InjectUtils;
 import io.openaev.utils.TargetType;
 import io.openaev.utils.fixtures.AssetGroupFixture;
